Comments 💬
Now I can create a comment, edit it and delete it. Just like I made a small change in the query code that is made to get the posts. Now when the query is made, in addition to fetching the posts, it will also fetch their respective comments and the authors of those comments.